refactor(api): move god to modules/ as controller, model, service
